refactor: readLineAsync 함수 input.js에서 분리
