+++ src/view/input.js
@@ -1,19 +1,5 @@
-import readline from "readline";
 import { RULE, VIEW_MESSAGE } from "../constants/index.js";
-
-const readLineAsync = (query = "") => {
-  return new Promise((resolve) => {
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-
-    rl.question(query, (input) => {
-      rl.close();
-      resolve(input);
-    });
-  });
-};
+import { readLineAsync } from "../utils/readLineAsync.js";
 
 export const retryUntilValid = async (getInputFn, validator) => {
   while (true) {
